# docs/contracts/openapi/generated.json
{
  "components": {
    "schemas": {
...
      "FieldError": {
        "additionalProperties": false,
        "description": "One field-level failure inside a problem response.\n\nPresent so that a rejected submission can be rendered against the form that produced\nit — FR-PLAT-11 requires deterministic failures to name the offending field rather\nthan only the operation.",
        "properties": {
          "code": {
            "description": "Stable machine code for this field.",
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Code",
            "type": "string"
          },
          "field": {
            "description": "Dotted path to the offending field, e.g. 'spec.family'.",
            "title": "Field",
            "type": "string"
          },
          "message": {
            "description": "Human-readable explanation.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "ProblemDetail": {
        "additionalProperties": false,
        "description": "An RFC 9457 problem response.\n\n`code` is the contract, not `title` or `status`: titles are prose and several\nconditions share a status, so a client that branches on anything else is branching on\nsomething the platform is free to reword.",
        "properties": {
          "code": {
            "description": "Stable machine-readable code, namespaced per module and enumerated in that module's spec. This is what clients branch on.",
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Explanation specific to this occurrence.",
            "title": "Detail"
          },
          "errors": {
            "default": [],
            "description": "Field-level errors, where the failure is field-attributable.",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Path of the failing request.",
            "title": "Instance"
          },
          "status": {
            "description": "HTTP status code.",
            "maximum": 599.0,
            "minimum": 100.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short human-readable summary, stable per type.",
            "title": "Title",
            "type": "string"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "OpenTelemetry trace id for this request. Appears in every log line for the request, so a support conversation starts with an identifier (FR-PLAT-42).",
            "title": "Trace Id"
          },
          "type": {
            "default": "about:blank",
            "description": "URI identifying the problem type; documentation lives at that URI.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "SettingCandidate": {
        "additionalProperties": false,
        "description": "What one layer offered, `None` when that layer says nothing.",
        "properties": {
          "source": {
            "$ref": "#/components/schemas/SettingSource"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "source"
        ],
        "title": "SettingCandidate",
        "type": "object"
      },
      "SettingResolution": {
        "additionalProperties": false,
        "description": "One setting's effective value and how it was arrived at (`07` §4.4).",
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/SettingCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "constraints": {
            "additionalProperties": true,
            "title": "Constraints",
            "type": "object"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "effective_value": {
            "title": "Effective Value"
          },
          "feature_flag": {
            "default": false,
            "description": "Feature flags gate optional or risk-bearing capabilities and default to the safe value (FR-PLAT-46).",
            "title": "Feature Flag",
            "type": "boolean"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "resolved_from": {
            "$ref": "#/components/schemas/SettingSource"
          },
          "type": {
            "$ref": "#/components/schemas/SettingType"
          }
        },
        "required": [
          "key",
          "effective_value",
          "resolved_from",
          "candidates",
          "type"
        ],
        "title": "SettingResolution",
        "type": "object"
      },
      "SettingSource": {
        "description": "Which layer supplied a value. Ordered by precedence, highest first.",
        "enum": [
          "env",
          "workspace",
          "default"
        ],
        "title": "SettingSource",
        "type": "string"
      },
      "SettingType": {
        "enum": [
          "bool",
          "int",
          "float",
          "string"
        ],
        "title": "SettingType",
        "type": "string"
      },
      "UpdateSettings": {
        "additionalProperties": false,
        "properties": {
          "values": {
            "additionalProperties": true,
            "description": "Setting key to value. Only declared keys are accepted; an unknown key is a 404 rather than a silently stored value nothing will ever read.",
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "values"
        ],
        "title": "UpdateSettings",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "GI Pricing Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "operationId": "get_settings_api_v1_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SettingResolution"
                  },
                  "title": "Response Get Settings Api V1 Settings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not permitted"
          }
        },
        "summary": "Effective settings with their resolution source",
        "tags": [
          "platform"
        ]
      },
      "put": {
        "description": "Validate, write and audit each change in one transaction (`06` R2, FR-PLAT-31).",
        "operationId": "update_settings_api_v1_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SettingResolution"
                  },
                  "title": "Response Update Settings Api V1 Settings Put",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not permitted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Update workspace settings",
        "tags": [
          "platform"
        ]
      }
    },
...
  }
}